refactor(core): Remove circular dependency on `render3` and `ng_module`
Extracted `NgModeDef` into a separate file to break the circular dependency.
