+++ packages/core/src/r3_symbols.ts
@@ -23,7 +23,8 @@
 
 export {ɵɵinject} from './di/injector_compatibility';
 export {ɵɵdefineInjectable, ɵɵdefineInjector, ɵɵInjectableDef, ɵɵInjectorDef} from './di/interface/defs';
-export {NgModuleDef, ɵɵNgModuleDefWithMeta} from './metadata/ng_module';
+export {ɵɵNgModuleDefWithMeta} from './metadata/ng_module';
+export {NgModuleDef} from './metadata/ng_module_def';
 export {ɵɵdefineNgModule} from './render3/definition';
 export {ɵɵFactoryDef} from './render3/interfaces/definition';
 export {setClassMetadata} from './render3/metadata';
