+++ packages/core/test/render3/ivy/jit_spec.ts
@@ -7,15 +7,17 @@
  */
 import 'reflect-metadata';
 
-import {ElementRef, QueryList, ɵɵsetComponentScope as setComponentScope} from '@angular/core';
+import {ElementRef, QueryList} from '@angular/core';
 import {Injectable} from '@angular/core/src/di/injectable';
 import {setCurrentInjector, ɵɵinject} from '@angular/core/src/di/injector_compatibility';
 import {ɵɵdefineInjectable, ɵɵInjectorDef} from '@angular/core/src/di/interface/defs';
 import {ivyEnabled} from '@angular/core/src/ivy_switch';
 import {ContentChild, ContentChildren, ViewChild, ViewChildren} from '@angular/core/src/metadata/di';
-import {Component, Directive, HostBinding, HostListener, Input, Output, Pipe} from '@angular/core/src/metadata/directives';
-import {NgModule, NgModuleDef} from '@angular/core/src/metadata/ng_module';
+import {Component, Directive, HostBinding, HostListener, Input, Pipe} from '@angular/core/src/metadata/directives';
+import {NgModule} from '@angular/core/src/metadata/ng_module';
+import {NgModuleDef} from '@angular/core/src/metadata/ng_module_def';
 import {ComponentDef, FactoryFn, PipeDef} from '@angular/core/src/render3/interfaces/definition';
+
 
 
 ivyEnabled && describe('render3 jit', () => {
